software/ccam/requirements.txt
pyserial
psutil
exifread
requests
picamera
gunicorn
